Add changeSecret function
Added ability to change secret character at which ordering is added to encryption key. Environmental variables cannot be updated mid-script, so right now there are 2 variables, secret and newSecret, and newSecret has to be manually updated before running changeSecret() with the new secret character. Also, right now this would only work for one key, but since all keys share secret, they need to all be updated at once (which isn't hard to do). This problem is not a real one, since saving things in a .env isn't how this will actually work in the end when information is distributed between parties.